+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Example to test openDir on /home/web/tmp folder.
 ### Changed
+- Prevent Zig openDir on /home/web/tmp folder.
 ### Fixed